move purity estimation to separate file and create new variable for clarity
