+++ cirq/experiments/purity_estimation_test.py
@@ -1,0 +1,25 @@
+# Copyright 2020 The Cirq Developers
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import numpy as np
+
+from cirq.experiments.cross_entropy_benchmarking import (
+    purity_from_probabilities)
+
+
+def test_purity_from_probabilities():
+    probabilities = np.random.uniform(0, 1, size=4)
+    probabilities /= np.sum(probabilities)
+    purity = purity_from_probabilities(4, probabilities)
+    np.testing.assert_allclose(purity, np.var(probabilities) * 80 / 3)
